BS-41: Cleaned up code
